put test python files in a flat folder
