+++ cmake/onnxruntime_python.cmake
@@ -173,6 +173,7 @@ endif()
 
 file(GLOB onnxruntime_python_test_srcs CONFIGURE_DEPENDS
     "${ONNXRUNTIME_ROOT}/test/python/*.py"
+    "${ORTTRAINING_SOURCE_DIR}/test/python/*.py"
 )
 file(GLOB onnxruntime_python_tools_srcs CONFIGURE_DEPENDS
     "${ONNXRUNTIME_ROOT}/python/tools/*.py"
