--- orttraining/orttraining/test/python/helpers/utils.py
@@ -1,36 +1,0 @@
-import torch
-
-def warmup_cosine(x, warmup=0.002):
-    if x < warmup:
-        return x/warmup
-    return 0.5 * (1.0 + torch.cos(math.pi * x))
-
-def warmup_constant(x, warmup=0.002):
-    if x < warmup:
-        return x/warmup
-    return 1.0
-
-def warmup_linear(x, warmup=0.002):
-    if x < warmup:
-        return x/warmup
-    return max((x - 1. )/ (warmup - 1.), 0.)
-    
-def warmup_poly(x, warmup=0.002, degree=0.5):
-    if x < warmup:
-        return x/warmup
-    return (1.0 - x)**degree
-
-
-SCHEDULES = {
-    'warmup_cosine':warmup_cosine,
-    'warmup_constant':warmup_constant,
-    'warmup_linear':warmup_linear,
-    'warmup_poly':warmup_poly,
-}
-
-def get_lr(args, training_steps, schedule='warmup_poly'):
-    if args.max_steps == -1:
-        return args.learning_rate
-
-    schedule_fct = SCHEDULES[schedule]
-    return args.learning_rate * schedule_fct(training_steps / args.max_steps, args.warmup_proportion)
