+++ orttraining/orttraining/test/python/orttraining_test_transformers.py
@@ -5,193 +5,16 @@ from __future__ import print_function
 import unittest
 import shutil
 import pytest
-import sys
 import os
-from enum import Enum
 
-# set transformer_repo to the huggingface repo location
-transformer_repo = ''
-nvidia_deep_learning_examples_repo = ''
+from transformers import (BertConfig, BertForPreTraining)
 
-sys.path.append(transformer_repo)
+from orttraining_test_data_loader import ids_tensor, BatchArgsOption
+from orttraining_test_utils import run_test, get_lr
 
-# from configuration_common_test import ConfigTester
+from onnxruntime.capi.ort_trainer import ORTTrainer, IODescription, ModelDescription, LossScaler
 
-# from modeling_common_test import (CommonTestCases, ids_tensor, floats_tensor)
-
-import random
-global_rng = random.Random()
-
-def ids_tensor(shape, vocab_size, rng=None, name=None):
-    """Creates a random int32 tensor of the shape within the vocab size."""
-    if rng is None:
-        rng = global_rng
-
-    total_dims = 1
-    for dim in shape:
-        total_dims *= dim
-
-    values = []
-    for _ in range(total_dims):
-        values.append(rng.randint(0, vocab_size - 1))
-
-    return torch.tensor(data=values, dtype=torch.long).view(shape).contiguous()
-
-
-def floats_tensor(shape, scale=1.0, rng=None, name=None):
-    """Creates a random float32 tensor of the shape within the vocab size."""
-    if rng is None:
-        rng = global_rng
-
-    total_dims = 1
-    for dim in shape:
-        total_dims *= dim
-
-    values = []
-    for _ in range(total_dims):
-        values.append(rng.random() * scale)
-
-    return torch.tensor(data=values, dtype=torch.float).view(shape).contiguous()
-
-
-from transformers import (BertConfig, BertModel, BertForMaskedLM,
-                            BertForNextSentencePrediction, BertForPreTraining,
-                            BertForQuestionAnswering, BertForSequenceClassification,
-                            BertForTokenClassification, BertForMultipleChoice)
-
-sys.path.append("/bert_ort/liqun/onnxruntime/build/Linux/Debug/")
-import onnxruntime as ort
-from onnxruntime.capi.ort_trainer import ORTTrainer, IODescription, ModelDescription, LossScaler, generate_sample
 import torch
-
-from helpers.bert_postprocess import postprocess_model
-
-# sys.path.append(os.path.join(nvidia_deep_learning_examples_repo, 'PyTorch/LanguageModeling/BERT'))
-# from run_pretraining import postprocess_model
-
-
-def map_optimizer_attributes(name):
-    no_decay_keys = ["bias", "gamma", "beta", "LayerNorm"]
-    no_decay = any(no_decay_key in name for no_decay_key in no_decay_keys)
-    if no_decay:
-        return {"alpha": 0.9, "beta": 0.999, "lambda": 0.0, "epsilon": 1e-6}
-    else:
-        return {"alpha": 0.9, "beta": 0.999, "lambda": 0.01, "epsilon": 1e-6}
-
-from torch.utils.data import Dataset, DataLoader
-class OrtTestDataset(Dataset):
-    def __init__(self, input_desc, seq_len, device):
-        import copy
-        self.input_desc_ = copy.deepcopy(input_desc)
-        for input_desc in self.input_desc_:
-            shape_ = []
-            for i, axis in enumerate(input_desc.shape_):
-                if axis == 'max_seq_len_in_batch':
-                    shape_ = shape_ + [seq_len, ]
-                elif axis != 'batch':
-                    shape_ = input_desc.shape_[i]
-            input_desc.shape_ = shape_
-        self.device_ = device
-
-    def __len__(self):
-        return 100
-
-    def __getitem__(self, item):
-        input_batch = []
-        for input_desc in self.input_desc_:
-            input_sample = generate_sample(input_desc, self.device_)
-            input_batch.append(input_sample)
-        return input_batch
-
-def create_ort_test_dataloader(input_desc, batch_size, seq_len, device):
-    dataset = OrtTestDataset(input_desc, seq_len, device)
-    return DataLoader(dataset, batch_size=batch_size)
-
-class BatchArgsOption(Enum):
-    List = 1
-    Dict = 2
-    ListAndDict = 3
-
-def split_batch(batch, input_desc, args_count):
-    total_argument_count = len(input_desc)
-                # batch=[input_ids[batch, seglen], attention_mask[batch, seglen], token_type_ids[batch,seglen], token_type_ids[batch, seglen]]
-    args = []   # (input_ids[batch, seglen], attention_mask[batch, seglen])
-    kwargs = {} # {'token_type_ids': token_type_ids[batch,seglen], 'position_ids': token_type_ids[batch, seglen]}
-    for i in range(args_count):
-        args = args + [batch[i]]
-
-    for i in range(args_count, total_argument_count):
-        kwargs[input_desc[i].name_] = batch[i]
-
-    return args, kwargs
-
-def run_test(model, model_desc, device, args, gradient_accumulation_steps, fp16,
-    allreduce_post_accumulation, get_lr_this_step, use_internal_get_lr_this_step, loss_scaler, use_internal_loss_scaler, 
-    batch_args_option):
-    dataloader = create_ort_test_dataloader(model_desc.inputs_, args.batch_size, args.seq_len, device)
-
-    model = ORTTrainer(model, None, model_desc, "LambOptimizer",
-        map_optimizer_attributes=map_optimizer_attributes,
-        learning_rate_description=IODescription('Learning_Rate', [1,], torch.float32),
-        device=device, postprocess_model=postprocess_model,
-        gradient_accumulation_steps=gradient_accumulation_steps,                
-        # BertLAMB default initial settings: b1=0.9, b2=0.999, e=1e-6
-        world_rank=args.local_rank, world_size=args.world_size,
-        use_mixed_precision=fp16,
-        allreduce_post_accumulation=allreduce_post_accumulation,
-        get_lr_this_step=get_lr_this_step if use_internal_get_lr_this_step else None,
-        loss_scaler=loss_scaler if use_internal_loss_scaler else None)
-
-    # trainig loop
-    eval_batch = None
-    model.train()
-    for step, batch in enumerate(dataloader):
-        if eval_batch is None:
-            eval_batch = batch
-
-        if not use_internal_get_lr_this_step:
-            lr = get_lr_this_step(step)
-            learning_rate = torch.tensor([lr])
-
-        if not use_internal_loss_scaler and fp16:
-            loss_scale = torch.tensor([loss_scaler.loss_scale_])
-
-        if batch_args_option == BatchArgsOption.List:
-            if not use_internal_get_lr_this_step:
-                batch = batch + [learning_rate, ]
-            if not use_internal_loss_scaler and fp16:
-                batch = batch + [loss_scale, ]
-            outputs = model(*batch)
-        elif batch_args_option == BatchArgsOption.Dict:
-            args, kwargs = split_batch(batch, model_desc.inputs_, 0)
-            if not use_internal_get_lr_this_step:
-                kwargs['Learning_Rate'] = learning_rate
-            if not use_internal_loss_scaler and fp16:
-                kwargs[model.loss_scale_input_name] = loss_scale
-            outputs = model(*args, **kwargs)
-        else:
-            args_count = int(len(model_desc.inputs_) / 2)   # approx helf args, half kwargs
-            args, kwargs = split_batch(batch, model_desc.inputs_, args_count)
-            if not use_internal_get_lr_this_step:
-                kwargs['Learning_Rate'] = learning_rate
-            if not use_internal_loss_scaler and fp16:
-                kwargs[model.loss_scale_input_name] = loss_scale
-            outputs = model(*args, **kwargs)
-
-    # eval
-    model.eval()
-    if batch_args_option == BatchArgsOption.List:
-        outputs = model(*batch)
-    elif batch_args_option == BatchArgsOption.Dict:
-        args, kwargs = split_batch(batch, model_desc.inputs_, 0)
-        outputs = model(*args, **kwargs)
-    else:
-        args_count = int(len(model_desc.inputs_) / 2)   # approx helf args, half kwargs
-        args, kwargs = split_batch(batch, model_desc.inputs_, args_count)
-        outputs = model(*args, **kwargs)
-
-    return (output.cpu().numpy() for output in outputs)
-
 
 class BertModelTest(unittest.TestCase):
 
@@ -334,7 +157,6 @@ class BertModelTest(unittest.TestCase):
                 "local_rank world_size max_steps learning_rate warmup_proportion batch_size seq_len")
             args = MyArgs(local_rank=0, world_size=1, max_steps=100, learning_rate=0.00001, warmup_proportion=0.01, batch_size=13, seq_len=7)
 
-            from helpers.utils import get_lr
             def get_lr_this_step(global_step):
                 return get_lr(args, global_step)
             loss_scaler = LossScaler('loss_scale_input_name', True, up_scale_window=2000)
@@ -345,7 +167,6 @@ class BertModelTest(unittest.TestCase):
             option_use_internal_get_lr_this_step = False
             option_use_internal_loss_scaler = True
             option_split_batch = [BatchArgsOption.ListAndDict]
-            # TODO: with with fetches
 
             for gradient_accumulation_steps in option_gradient_accumulation_steps:
                 for fp16 in option_fp16:
