+++ orttraining/pytorch_frontend_examples/mnist_training.py
@@ -19,14 +19,6 @@ from torchvision import datasets, transforms
 import numpy as np
 import os
 
-# TODO: remove after ready for CV
-# import sys
-# sys.path.insert(0, '/bert_ort/liqun/onnxruntime/build/Linux/Debug/')
-# import onnxruntime as ort
-
-# sys.path.insert(0, '/bert_ort/liqun/onnxruntime/onnxruntime/python/')
-# from ort_trainer import IODescription, ModelDescription, ORTTrainer, ORTModel
-
 from onnxruntime.capi.ort_trainer import IODescription, ModelDescription, ORTTrainer, ORTModel
 from mpi4py import MPI
 from onnxruntime.capi._pybind_state import set_cuda_device_id
